Pulled LedgerEventRow component into its own files, fixed references
